Move WTF::Locker specializations to Locker.h https://bugs.webkit.org/show_bug.cgi?id=302834 rdar://165097267
Reviewed by Chris Dumez.

It took me an embarrassingly long time to realize that the primary Locker<T>
class is not the one in Locker.h, but rather a separate template specialization
in Lock.h.

Moved the specialization to Locker.h.

* Source/WTF/wtf/Lock.h:
* Source/WTF/wtf/Locker.h:

Canonical link: https://commits.webkit.org/303303@main
